+++ src/widgets/transformCameraByObj.js
@@ -1,4 +1,5 @@
 const PhSim = require("../index");
+const Widget = require("../widget");
 
 /**
  * 
@@ -11,24 +12,48 @@ const PhSim = require("../index");
 
 
 PhSim.Widgets.transformCameraByObj = function(dyn_object) {
+
+    var w = new Widget(dyn_object);
+
     var self = this;
 
     var dx;
     var dy;
 
-    this.on("beforeupdate",function(){
-        dx = dyn_object.matter.position.x;
-        dy = dyn_object.matter.position.y;
-    },{
+    // beforeUpdate
+
+    var beforeUpdate = function(){
+        if(w.status === "enabled") {
+            dx = dyn_object.matter.position.x;
+            dy = dyn_object.matter.position.y;
+        }
+    }
+
+    this.on("beforeupdate",beforeUpdate,{
         "slEvent": true
     });
 
-    this.on("afterupdate",function(){
-        dx = dyn_object.matter.position.x - dx;
-        dy = dyn_object.matter.position.y - dy;
-        self.camera.translate(-dx,-dy);
-    },{
+    // afterupdate
+
+    var afterUpdate = function(){
+        if(w.status === "enabled") {
+            dx = dyn_object.matter.position.x - dx;
+            dy = dyn_object.matter.position.y - dy;
+            self.camera.translate(-dx,-dy);
+        }
+    }
+
+    this.on("afterupdate",afterUpdate,{
         "slEvent": true
     });
+
+    // Widget Destruction
+
+    w.on("destoy",function(){
+        self.off("afterupdate",afterUpdate);
+        self.off("beforeupdate",beforeUpdate);
+    });
+
+    return w;
 
 }