+++ src/events/eventStack.js
@@ -164,6 +164,8 @@ const EventStack = function() {
 
 	this.firstslupdate = [];
 
+	this.beforefirstslupdate = [];
+
 	/** Array of functions to be executed before the simulation exit */
 
 	this.exit = []
@@ -197,6 +199,12 @@ const EventStack = function() {
 	 */
 
 	this.wfunctionerror = [];
+
+	/**
+	 * Array of functions to be executed after the canvas are cleared.
+	 */
+
+	this.aftercanvasclear = [];
 
 
 }