+++ src/loop/update.js
@@ -84,7 +84,7 @@ PhSim.prototype.loopFunction = function() {
 		this.callEventClass("beforeupdate",this,beforeUpdateEvent);
 
 		if(!this.firstSlUpdate) {
-			this.callEventClass("beforefirstslupdate",this,afterUpdateEvent);
+			this.callEventClass("beforefirstslupdate",this,new PhSim.Events.PhSimDynEvent());
 		}
 
 		this.updateDate = new Date();
@@ -127,6 +127,8 @@ PhSim.prototype.loopFunction = function() {
 				this.ctx.fillRect(0 - this.camera.x,0 - this.camera.y,this.width / this.camera.scale,this.height / this.camera.scale);
 			}
 		}
+
+		this.callEventClass("aftercanvasclear",this,new PhSim.Events.PhSimDynEvent());
 
 		for(let i = 0; i < this.objUniverse.length; i++) {
 			this.updateDynObj(this.objUniverse[i]);