+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.3.0-alpha (Unreleased)
 
+* Added ImageData, Canvas and HTMLImageElement support for PhRender.renderSpriteByCenter. 
+
+* Fixed `onfirstslupdate` event bug that makes it not run at all.
+
 * Added `aftercanvasclear` event.
 
 * `PhSim.Vector.neg` has been added. This creates a vector that is equal in length but opposite in direction of the original vector.